[REFACTOR] Resolving Security, Maintainability and Reliability code smells
