+++ src/main/java/ppl/momofin/momofinbackend/repository/UserRepository.java
@@ -3,8 +3,6 @@ package ppl.momofin.momofinbackend.repository;
 import ppl.momofin.momofinbackend.model.Organization;
 import ppl.momofin.momofinbackend.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
