+++ src/main/java/ppl/momofin/momofinbackend/response/AuthResponse.java
@@ -1,0 +1,4 @@
+package ppl.momofin.momofinbackend.response;
+
+public interface AuthResponse {
+}
